feat(AIModels): refactor AI credential and model forms into separate components for improved modularity and maintainability
